% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_adm1.R
\docType{data}
\name{est_adm1}
\alias{est_adm1}
\title{Estonia county borders}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 15 rows and 11 columns.}
\usage{
est_adm1
}
\description{
Estonia ADM1 data from GADM
}
\keyword{datasets}
