% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcshapes-package.R
\docType{package}
\name{wcshapes-package}
\alias{wcshapes}
\alias{wcshapes-package}
\title{wcshapes: Compute Spatial Lags For State And Panel Data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Andreas Beger \email{adbeger@gmail.com} (0000-0003-1883-3169)

}
\keyword{internal}
